separate query bus to separate file
