+++ PL.Authorization.Infrastructure/CommandBus.cs
@@ -21,18 +21,4 @@ namespace PL.Authorization.Infrastructure
             await _mediator.Publish(command);
         }
     }
-
-    internal class QueryBus : IQueryBus
-    {
-        private readonly IMediator _mediator;
-
-        public QueryBus(IMediator mediator)
-        {
-            _mediator = mediator;
-        }
-        public async Task<TResult> ExecuteQueryAsync<TResult>(IQuery<TResult> query)
-        {
-            return await _mediator.Send(query);
-        }
-    }
 }
